# api/openapi.yaml
openapi: 3.1.0
info:
  title: Windlass API
  description: |
    HTTP API of the Windlass Docker Compose control plane. All endpoints are
    versioned under /api/v1. Authentication uses an HTTP-only session cookie
    obtained from /auth/login or /auth/setup.
  version: 0.1.0
servers:
  - url: /api/v1
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: windlass_session
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
...
paths:
...
  /webhooks/{provider}/{project}:
    post:
      summary: Git push webhook (public, signature-verified)
      operationId: webhook
      parameters:
        - { name: provider, in: path, required: true, schema: { type: string, enum: [github, gitlab] } }
        - { name: project, in: path, required: true, schema: { type: string } }
      responses:
        "201": { description: Deployment triggered }
        "200": { description: Ignored (branch mismatch or auto-deploy off) }
        "202": { description: Skipped, a deployment is already running }
        "403": { description: Signature verification failed }
...
  /projects/{name}/deployments:
    get:
      summary: List deployments
      operationId: listDeployments
      security: [{ session: [] }]
      responses:
        "200": { description: Array of deployments, newest first }
    post:
      summary: Deploy (member)
      operationId: createDeployment
      security: [{ session: [] }]
      responses:
        "201": { description: Deployment queued }
        "409": { description: A deployment is already running }
  /projects/{name}/deployments/{number}:
    get:
      summary: Get a deployment
      operationId: getDeployment
      security: [{ session: [] }]
      responses:
        "200": { description: Deployment }
  /projects/{name}/deployments/{number}/events:
    get:
      summary: Deployment event stream (SSE, replay + live, Last-Event-ID)
      operationId: deploymentEvents
      security: [{ session: [] }]
      responses:
        "200": { description: text/event-stream; closes after terminal replay }
  /projects/{name}/deployments/{number}/rollback:
    post:
      summary: Roll back to this deployment's recorded image digests (member)
      operationId: rollback
      security: [{ session: [] }]
      responses:
        "201": { description: Rollback deployment queued }
...
  /git/connections:
    get:
      summary: List git connections (admin)
      operationId: listGitConnections
      security: [{ session: [] }]
      responses:
        "200": { description: Array of connections (tokens never returned) }
    post:
      summary: Add a git connection (admin; token stored encrypted)
      operationId: createGitConnection
      security: [{ session: [] }]
      responses:
        "201": { description: Created }
  /git/connections/{id}:
    delete:
      summary: Delete a git connection (admin)
      operationId: deleteGitConnection
      security: [{ session: [] }]
      responses:
        "204": { description: Deleted }
  /git/connections/github/connect:
    get:
      summary: Begin one-click GitHub connect using the instance's OAuth app with repo scope (admin)
      operationId: gitConnectStart
      security: [{ session: [] }]
      responses:
        "302": { description: Redirect to GitHub authorize }
        "400":
          description: GitHub OAuth app not configured
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...